Rules for Contributions
Unlike the Code of Conduct the Rules for Contributions is a separate document that allows an individual or agent to contribute to an organization that is built on transparency and accountability. This is why we are choosing to be open sourced and will remain open sourced until we are no longer needed to provide our services to the public at large.